fix delete transaction

+++ src/controllers/transactions.js
@@ -56,10 +56,14 @@ export async function deleteTransaction(req, res){
     try{
         const session = await db.collection("sessions").findOne({token});
         if(!session) return res.status(401).send("Erro de autenticação");
-        const deleted = await db.collection('transactions').deleteOne({_id: new ObjectId(id)});
+        const deleted = await db.collection('transactions').deleteOne({_id: new ObjectId(id), userID: session.userId});
         if(deleted.deletedCount === 0) return res.sendStatus(404);
         return res.sendStatus(204);
     }catch(err){
         return res.status(500).send("Erro interno do servidor");
     }
+}
+
+export async function updateTransaction(req, res){
+
 }
